tests/wpt/web-platform-tests/service-workers/service-worker/clients-matchall-include-uncontrolled.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.matchAll with includeUncontrolled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
const base_url = 'resources/blank.html'; // This is out-of-scope.
const scope = base_url + '?clients-matchAll-includeUncontrolled';
let frames = [];

// Creates 3 iframes, 2 for in-scope and 1 for out-of-scope. Returns the frame
// opened for scope + '#2'.
async function create_iframes(scope) {
  frames.push(await with_iframe(base_url));
  frames.push(await with_iframe(scope + '#1'));
  frames.push(await with_iframe(scope + '#2'));
  return frames[2];
}

const expected_without_include_uncontrolled = [
    // visibilityState, focused, url, type, frameType
    ['visible', false, new URL(scope + '#1', location).toString(), 'window', 'nested'],
    ['visible', true, new URL(scope + '#2', location).toString(), 'window', 'nested']
];

const expected_with_include_uncontrolled = [
    // visibilityState, focused, url, type, frameType
    ['visible', true, location.href, 'window', 'top-level'],
    ['visible', false, new URL(scope + '#1', location).toString(), 'window', 'nested'],
    ['visible', true, new URL(scope + '#2', location).toString(), 'window', 'nested'],
    ['visible', false, new URL(base_url, location).toString(), 'window', 'nested']
];

function test_matchall(frame, expected, query_options) {
  // Make sure we have focus for '#2' frame and its parent window.
  frame.focus();
  frame.contentWindow.focus();
  expected.sort((a, b) => a[2] > b[2] ? 1 : -1);
  return new Promise((resolve, reject) => {
    const channel = new MessageChannel();
    channel.port1.onmessage = e => {
      // Ignore hidden clients which may be coming from background tabs, or
      // clients unrelated to this test.
      const data = e.data.filter(info => {
        return info[0] == 'visible' &&
               info[2].indexOf('service-worker') > -1;
      });
      data.sort((a, b) => a[2] > b[2] ? 1 : -1);
      assert_equals(data.length, expected.length);
      for (let i = 0; i < data.length; i++)
        assert_array_equals(data[i], expected[i]);
      resolve();
    };
    frame.contentWindow.navigator.serviceWorker.controller.postMessage(
        {port:channel.port2, options:query_options},
        [channel.port2]);
  });
}

// Run clients.matchAll without and with includeUncontrolled=true.
// (We want to run the two tests sequentially in the same promise_test
// so that we can use the same set of iframes without intefering each other.
promise_test(async t => {
  const registration =
      await service_worker_unregister_and_register(
          t, 'resources/clients-matchall-worker.js', scope);
  t.add_cleanup(() => service_worker_unregister(t, scope));
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await create_iframes(scope);
  await test_matchall(frame, expected_without_include_uncontrolled);
  await test_matchall(frame, expected_with_include_uncontrolled,
                      {includeUncontrolled:true});
}, 'Verify matchAll() respect includeUncontrolled');
</script>
